refactor(usage): split dashboard.ts into responsibility modules
